requirements.txt
python-twitter==3.5
mwclient==0.10.1
mwparserfromhell==0.5.4
Pillow==8.0.1
schedule==0.6.0
nltk==3.5